route-53/api-gw-alias/requirements.txt
attrs==21.4.0
aws-cdk-lib==2.27.0
cattrs==22.1.0
constructs==10.1.33
exceptiongroup==1.0.0rc8
jsii==1.60.1
publication==0.0.3
python-dateutil==2.8.2
python-dotenv==0.20.0
six==1.16.0
typing-extensions==4.2.0
